app/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Google ADK and AI
google-adk==1.2.1
google-cloud-aiplatform[adk]==1.88.0
google-generativeai==0.8.3

# Google APIs
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0

# Data handling
pandas==2.1.4
pydantic==2.5.2
pydantic-settings==2.1.0

# WebSocket support
websockets==12.0

# Audio processing
numpy==1.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0